requirements.txt
azure-core==1.30.2
azure-storage-blob==12.21.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
idna==3.7
isodate==0.6.1
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.2.2
